--- oidc/jwks_url.go
@@ -1,5 +1,0 @@
-package oidc
-
-var JWKsUrls = map[string]string{
-	"google": "https://www.googleapis.com/oauth2/v1/certs",
-}
